Use dummy test app
