+++ test/dummy/config/routes.rb
@@ -11,4 +11,5 @@ Rails.application.routes.draw do
 
   # Defines the root path route ("/")
   # root "posts#index"
+  get ":controller(/:action)"
 end
