+++ test/views/jsx/admin/index.json.props
