+++ test/views/jsx/bad_pair.html.erb
