+++ test/views/jsx/bad_pair.json.props
