+++ test/views/jsx/bad_single.html.erb
