+++ test/views/jsx/simple.html.erb
