+++ test/views/jsx/simple.json.props
