+++ test/views/jsx/single_props_only.json.props
