+++ test/views/jsx/valid_pair.json.props
