+++ test/views/layouts/jsx_application.html.erb
